SUD-00010 Merge from 3.7.0 forward curve.
git-svn-id: http://fdevstorage/svn/IVRMstandard/trunk@476 2ab15184-0dcf-aa4b-bfde-16869b366a71

+++ EgarTransport/Transport/Host.cpp
@@ -579,7 +579,7 @@ STDMETHODIMP CHost::get_Statistic(IHostStatisticData** ppStatistic)
 
 			FILETIME ftUTC;
 			ftUTC.dwLowDateTime = (DWORD)llLogonTime;
-			ftUTC.dwHighDateTime = llLogonTime >> 32;
+			ftUTC.dwHighDateTime = static_cast<DWORD>(llLogonTime >> 32);
 			SYSTEMTIME sTime;
 			if(!FileTimeToSystemTime(&ftUTC,&sTime))
 				throw Error(L"FileTimeToSystemTime failed.", IID_IHost, E_FAIL);
